refactor(ts-sdk): clean up wallet types and move mocks to testing subpath
