+++ packages/babylon-ts-sdk/src/shared/index.ts
@@ -1,19 +1,12 @@
 // Shared utilities and cross-cutting concerns
 // This module contains wallet interfaces, errors, and common utilities
 
-// Wallet interfaces and implementations
+// Wallet interfaces
+export { BitcoinNetworks } from "./wallets";
 export type {
   BitcoinNetwork,
   BitcoinWallet,
   Hash,
+  SignInputOptions,
   SignPsbtOptions,
-} from "./wallets";
-
-export { createTaprootScriptPathSignOptions } from "./wallets";
-
-export { MockBitcoinWallet, MockEthereumWallet } from "./wallets";
-
-export type {
-  MockBitcoinWalletConfig,
-  MockEthereumWalletConfig,
 } from "./wallets";
