+++ packages/babylon-ts-sdk/src/tbv/core/utils/index.ts
@@ -14,3 +14,4 @@ export * from "./fee";
 export * from "./utxo";
 export * from "./transaction";
 export * from "./btc";
+export * from "./signing";
